fix: 타입 파일 명 변경
