--- src/@types/Address.d.ts
@@ -1,8 +1,0 @@
-export interface AddresDetail {
-  name: string,
-  phoneNumber: string,
-  zonecode: string,
-  address: string,
-  addressDetail: string,
-  defaultAddress: string,
-}
