+++ src/@types/Address.ts
@@ -1,0 +1,8 @@
+export interface IAddressDetail {
+  name: string;
+  phoneNumber: string;
+  zoneCode: string;
+  address: string;
+  addressDetail: string;
+  defaultAddress: string;
+}
