--- src/@types/Notification.d.ts
@@ -1,14 +1,0 @@
-declare module 'Notification' {
-  export interface IRealTimeNotification {
-    notificationId: number;
-    message: string;
-    type: string;
-    auctionId?: number;
-  }
-
-  export interface INotification extends IRealTimeNotification {
-    imageUrl: string;
-    isRead: boolean;
-    createdAt: string;
-  }
-}
