+++ src/@types/Notification.ts
@@ -1,0 +1,12 @@
+export interface IRealTimeNotification {
+  notificationId: number;
+  message: string;
+  type: string;
+  auctionId?: number;
+}
+
+export interface INotification extends IRealTimeNotification {
+  imageUrl: string;
+  isRead: boolean;
+  createdAt: string;
+}
