--- src/@types/Register.d.ts
@@ -1,10 +1,0 @@
-declare module 'Register' {
-  export interface IRegister {
-    productName: string;
-    description: string;
-    minPrice: number;
-    auctionRegisterType?: string;
-    category: string;
-    imageSequence?: { [k: string]: number };
-  }
-}
