+++ src/@types/Register.ts
@@ -1,0 +1,8 @@
+export interface IRegister {
+  productName: string;
+  description: string;
+  minPrice: number;
+  auctionRegisterType?: string;
+  category: string;
+  imageSequence?: { [k: string]: number };
+}
