--- src/@types/bid.d.ts
@@ -1,14 +1,0 @@
-declare module 'Bid' {
-  export interface IBidPostData {
-    auctionId: number;
-    bidAmount: number;
-  }
-
-  export interface IBidder {
-    isWinningBidder: boolean;
-    bidderNickname: string;
-    bidAmount: number;
-  }
-
-  export interface IBidderList extends PaginationData<IBidder> {}
-}
