+++ src/@types/bid.ts
@@ -1,0 +1,14 @@
+import { PaginationData } from './AuctionList';
+
+export interface IBidPostData {
+  auctionId: number;
+  bidAmount: number;
+}
+
+export interface IBidder {
+  isWinningBidder: boolean;
+  bidderNickname: string;
+  bidAmount: number;
+}
+
+export interface IBidderList extends PaginationData<IBidder> {}
